+++ addons/it/ticket/org_test.go
@@ -1,9 +1,9 @@
-package adecf_test
+package ticket_test
 
 import (
 	"testing"
 
-	"github.com/invopop/gobl/addons/it/adecf"
+	"github.com/invopop/gobl/addons/it/ticket"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 	"github.com/stretchr/testify/assert"
@@ -12,7 +12,7 @@ import (
 
 func TestNormalizeOrgItem(t *testing.T) {
 	t.Run("nil", func(t *testing.T) {
-		ad := tax.AddonForKey(adecf.V1)
+		ad := tax.AddonForKey(ticket.V1)
 		var item *org.Item
 		assert.NotPanics(t, func() {
 			ad.Normalizer(item)
@@ -21,6 +21,6 @@ func TestNormalizeOrgItem(t *testing.T) {
 	t.Run("with standard invoice", func(t *testing.T) {
 		inv := exampleStandardInvoice(t)
 		require.NoError(t, inv.Calculate())
-		assert.Equal(t, "services", inv.Lines[0].Item.Ext[adecf.ExtKeyProduct].String())
+		assert.Equal(t, "services", inv.Lines[0].Item.Ext[ticket.ExtKeyProduct].String())
 	})
 }