+++ addons/it/ticket/ticket.go
@@ -1,6 +1,6 @@
-// Package adecf handles the validation rules in order to use
+// Package ticket handles the validation rules in order to use
 // GOBL with the Italian Agenzia delle Entrate format.
-package adecf
+package ticket
 
 import (
 	"github.com/invopop/gobl/bill"
@@ -10,10 +10,10 @@ import (
 	"github.com/invopop/gobl/tax"
 )
 
-// Key to identify the AdE adecf addon
+// Key to identify the AdE ticket addon
 const (
 	// V1 for AdE format
-	V1 cbc.Key = "it-adecf-v1"
+	V1 cbc.Key = "it-ticket-v1"
 )
 
 func init() {
@@ -22,13 +22,24 @@ func init() {
 
 // This validation follows the rules of the Italian Agenzia delle Entrate
 // This addon will then be used to create documents using the following services
-// https://www.agenziaentrate.gov.it/portale/servizi/servizitrasversali/altri/cassetto-fiscale
 // https://www.agenziaentrate.gov.it/portale/schede/comunicazioni/fatture-e-corrispettivi
 func newAddon() *tax.AddonDef {
 	return &tax.AddonDef{
 		Key: V1,
 		Name: i18n.String{
-			i18n.EN: "Italy AdE adecf v1.x",
+			i18n.EN: "Italy AdE ticket v1.x",
+		},
+		Sources: []*cbc.Source{
+			{
+				Title:       i18n.NewString("Italian AdE Cassetto Fiscale"),
+				URL:         "https://www.agenziaentrate.gov.it/portale/schede/comunicazioni/fatture-e-corrispettivi",
+				ContentType: "application/pdf",
+			},
+			{
+				Title:       i18n.NewString("Italian AdE Fattura e Corrispettivi"),
+				URL:         "https://www.agenziaentrate.gov.it/portale/schede/comunicazioni/fatture-e-corrispettivi",
+				ContentType: "application/pdf",
+			},
 		},
 		Extensions: extensions,
 		Validator:  validate,
@@ -37,8 +48,9 @@ func newAddon() *tax.AddonDef {
 }
 
 func normalize(doc any) {
-	// nothing to normalize yet
 	switch obj := doc.(type) {
+	case *bill.Invoice:
+		normalizeInvoice(obj)
 	case *org.Item:
 		normalizeOrgItem(obj)
 	}